u-boot 1.1.1 now default for CDK
!!!some changes to previous behaviour, please read!!!

The make_xyz target-stuff is gone (not practical).

Use "make uboot" for everything. Configuration scheme is now like this:

In /boot/tuxbox-config there is a set of predefined configs with a
symbolic link "u-boot.config" which points to the desired configuration.

If the link does not exist when "make" is called it defaults to the CDK-config,
otherwise it is not touched.

I think this solution is the best, you can even build a u-boot for a
different board now, isn't that great? ;)

Maybe someone else feels the need to extend the flashrules so that one could "make"
the desired u-boot-config without manually changing the symlink but I don't know
if that's really necessary.

Please report all pecularities to the ML or board.